confd: Dependencies, RPC socket permissions
- Support fetching SQLite C++ wrapper automatically

- Add an option to set the permission mask of the created RPC socket to
  a user-defined value

+++ recipes-custom/sativa-confd/files/src/daemon/DataStore.cpp
@@ -393,7 +393,7 @@ void DataStore::insertKey(const std::string_view &keyName, const PropertyValue &
     SQLite::Statement infoStmt(*this->db, "INSERT INTO PropertyKeys (key, valueType) "
             "VALUES (:key, :type);");
     infoStmt.bind(":key", keyName.data());
-    infoStmt.bind(":type", static_cast<uintptr_t>(type));
+    infoStmt.bind(":type", static_cast<long>(type));
 
     err = infoStmt.exec();
     if(!err) {
